+++ fuel-tx/src/transaction/types/script.rs
@@ -182,11 +182,11 @@ impl FormatValidityChecks for Script {
             consensus_params.base_asset_id(),
         )?;
         let script_params = consensus_params.script_params();
-        if self.script.len() > script_params.max_script_length as usize {
+        if self.script.len() as u64 > script_params.max_script_length {
             Err(CheckError::TransactionScriptLength)?;
         }
 
-        if self.script_data.len() > script_params.max_script_data_length as usize {
+        if self.script_data.len() as u64 > script_params.max_script_data_length {
             Err(CheckError::TransactionScriptDataLength)?;
         }
